# Lint/GatewayService/src/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "FACEOPEN - API",
        "description": "All api description",
        "contact": {
            "name": "Ravi Ranjan, Rishabh Sonal , Akhil Regonda",
            "email": "ravi.ranjan@techolution.com, rishabh@techolution.com, akhil.regonda@techolution.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "host": "localhost:3500",
    "basePath": "/api",
...
    "schemes": [
        "https"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
...
        "/user/registration/validate/token": {
            "post": {
                "tags": [
                    "Registration"
                ],
                "description": "User Validate Token",
                "parameters": [
                    {
                        "name": "Validate Token",
                        "in": "body",
                        "description": "Update validate token",
                        "schema": {
                            "$ref": "#/definitions/validate_token"
                        }
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Token Validated Successfully",
                        "schema": {
                            "$ref": "#/definitions/validate_token_response"
                        }
                    }
                }
            }
        },
        "/user/registration/blurdetect": {
            "post": {
                "tags": [
                    "Registration"
                ],
                "description": "User Recognition Successfully",
                "parameters": [
                    {
                        "name": "Recognition",
                        "in": "body",
                        "description": "Update Recognition",
                        "schema": {
                            "$ref": "#/definitions/blurdetect"
                        }
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "User Recognition Successfully",
                        "schema": {
                            "$ref": "#/definitions/Generic_Response"
                        }
                    }
                }
            }
        },
        "/user/registration/add-user": {
            "post": {
                "tags": [
                    "Registration"
                ],
                "description": "User Registration",
                "parameters": [
                    {
                        "name": "Registration",
                        "in": "body",
                        "description": "Update Registration",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "User Registration Successfully",
                        "schema": {
                            "$ref": "#/definitions/Generic_Response"
                        }
                    }
                }
            }
        },
        "/user/registration/ratings": {
            "post": {
                "tags": [
                    "Registration"
                ],
                "description": "User Rating Successfully",
                "parameters": [
                    {
                        "name": "Rating",
                        "in": "body",
                        "description": "Update Rating",
                        "schema": {
                            "$ref": "#/definitions/Rating"
                        }
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "User Rating Successfully",
                        "schema": {
                            "$ref": "#/definitions/Generic_Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "User": {
            "properties": {
                "access": {
                    "type": "object"
                },
                "accessPoints": {
                    "type": "object"
                },
                "email": {
                    "type": "string"
                },
                "employee_id": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "phoneNumber": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                },
                "name": {
                    "type": "object"
                },
                "otherDetails": {
                    "type": "object"
                }
            }
        },
        "Generic_Response": {
            "properties": {
                "status": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "validate_token": {
            "properties": {
                "deviceInfo": {
                    "type": "string"
                },
                "onPrem_id": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                }
            }
        },
        "validate_token_response": {
            "properties": {
                "status": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string"
                },
                "lastName": {
                    "type": "string"
                },
                "dob": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "iat": {
                    "type": "string"
                },
                "rejected": {
                    "type": "string"
                }
            }
        },
        "blurdetect": {
            "properties": {
                "image": {
                    "type": "string"
                }
            }
        },
        "Rating": {
            "properties": {
                "rating": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "log": {
                    "type": "string"
                },
                "_id": {
                    "type": "string"
                },
                "comment": {
                    "type": "string"
                }
            }
        }
    }
}